--- Assets/Hero Knight - Pixel Art/Demo/HeroKnight.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0f5034bc83802944c9bca57c55e08978
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
