--- Assets/The_Hunt/Scenes/Loading.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c1c429c210c155a4d9decc755b16dd45
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
