--- Assets/The_Hunt/Scenes/Lose.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 31c3f0c2d5a1d8f40b8fff2f06455ec7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
