+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,28 +5,19 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 0
-    path: Assets/Scenes/MenuLevel.unity
-    guid: 2fa55462fc719a848b33dd6097f67562
-  - enabled: 0
-    path: Assets/Pixel_Quest/Scenes/TestLevel.unity
-    guid: 2cda990e2423bbf4892e6590ba056729
-  - enabled: 0
-    path: Assets/Scenes/Level_2.unity
-    guid: f9555ef8cf6b50447922872cea2af702
   - enabled: 1
-    path: Assets/Geo_Quest/Scenes/Scene_1.unity
-    guid: 8fcbbf205b9dfa043a98c17f73b8a9f0
+    path: Assets/The_Hunt/Scenes/Level_1.unity
+    guid: 998c318f327d97a458dc1021c9b8277f
   - enabled: 1
-    path: Assets/Geo_Quest/Scenes/Scene_2.unity
-    guid: 260432ea29022344188f6336eedb0f8a
+    path: Assets/The_Hunt/Scenes/Level_2.unity
+    guid: 6c25003690a43394cbf25be86f2f26d5
   - enabled: 1
-    path: Assets/Geo_Quest/Scenes/Scene_3.unity
-    guid: a3e1306a032fa06409935d9b7307ebf0
+    path: Assets/The_Hunt/Scenes/Level_3.unity
+    guid: b06adbd492734464baf5e7287671f9c0
   - enabled: 1
-    path: Assets/Geo_Quest/Scenes/Scene_4.unity
-    guid: 68b4ecac160239e4e9960e2bd97ef091
+    path: Assets/The_Hunt/Scenes/Level_4.unity
+    guid: 84fc1821d873bf645a116402bcb546ec
   - enabled: 1
-    path: Assets/Geo_Quest/Scenes/Scene_5.unity
-    guid: 54bbbdd43bd36f342a3654a849084ce6
+    path: Assets/The_Hunt/Scenes/Win.unity
+    guid: 75a3051a5039bff4e9b84a15b83edbb6
   m_configObjects: {}
